need repeatSync=true for testEditBooks.scr script
because it needs to sync method calls back and forth for elements
of the repeat tags. Those types get created symmetrically in the client
and server for this test so we need to sync on them but not send over
the create messages.

Previously that was on by default, but the registerSync call though was adding overhead
and breaking a case where we were needing the newX calls to be synchronized
(i.e. where we sync the repeat tags themselves using a separate property and
return the repeatTags from createElement rather than creating a new instance)